# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /preference:
    post:
      operationId: addProcessPartnerPreference
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessPreferences'
      - name: Authorization
        in: header
        type: string
        required: true

      responses:
        # Response code
        200:
          description: Partner Preference added
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      operationId: updateProcessPartnerPreference
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        schema:
          $ref: '#/definitions/ProcessPreferences'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference updated
          schema:
            $ref: '#/definitions/ApiResponse'
  /preference/{partnerID}:
    get:
      operationId: getProcessPartnerPreference
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference obtained
          schema:
            $ref: '#/definitions/ProcessPreferences'
    delete:
      operationId: deleteProcessPartnerPreference
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Partner Preference deleted
          schema:
            $ref: '#/definitions/ApiResponse'
...
  /document:
    post:
      operationId: addDocumentMetadata
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessDocumentMetadata'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document added
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      operationId: updateDocumentMetadata
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        schema:
          $ref: '#/definitions/ProcessDocumentMetadata'
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document updated
          schema:
            $ref: '#/definitions/ApiResponse'
  /document/{documentID}:
    delete:
      operationId: deleteDocument
      produces:
      - application/json
      parameters:
      - in: path
        name: documentID
        type: string
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document deleted
          schema:
            $ref: '#/definitions/ApiResponse'
  /document/{partnerID}/{type}/{source}/{status}:
    get:
      operationId: getDocuments
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - in: path
        name: type
        type: string
        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
        required: true
      - in: path
        name: status
        type: string
        description: The status of the document. (APPROVED, WAITINGRESPONSE, DENIED or PASTDUE)
        required: true
      - in: path
        name: source
        type: string
        description: The source of the document (whether it is sent or received by the partner - SENT, RECEIVED).
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document obtained
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessDocumentMetadata'
  /document/{partnerID}/{type}/{source}:
    get:
      operationId: getDocuments
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - in: path
        name: type
        type: string
        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
        required: true
      - in: path
        name: source
        type: string
        description: The source of the document (whether it is sent or received by the partner - SENT, RECEIVED).
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document obtained
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessDocumentMetadata'
  /document/{partnerID}/{type}:
    get:
      operationId: getDocuments
      produces:
      - application/json
      parameters:
      - in: path
        name: partnerID
        type: string
        description: The id of the partner.
        required: true
      - in: path
        name: type
        type: string
        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
        required: true
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Document obtained
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessDocumentMetadata'
...
definitions:
...
  ProcessPreferences:
    title: Business Process Preferences Object of a partner for all other partners
    type: object
    properties:
      partnerID:
        type: string
      preferences:
        type: array
        items:
          $ref: '#/definitions/ProcessPreference'
  ProcessPreference:
    title: Business Process Preference Object for a specific case (e.g. default or for a partner)
    type: object
    properties:
      targetPartnerID:
        type: string
        description: The value is DEFAULT or the partner ID for whom this preference is specified
      processOrder:
        type: array
        items:
          type: string
          enum:
          - CATALOGUE
          - NEGOTIATION
          - ORDER
          - REMITTANCEADVICE
          - INVOICE
          - TRACKING
          - FULFILMENT
          - TRANSPORT_EXECUTION_PLAN
          - ITEM_INFORMATION_REQUEST
          - PRODUCTCONFIGURATION
          - PPAP
          - OTHER
...
  ProcessDocumentMetadata:
    title: Business Process Document Metadata Object
    type: object
    properties:
      documentID:
        type: string
      submissionDate:
        type: string
      type:
        type: string
        enum:
        - CATALOGUE
        - REQUESTFORQUOTATION
        - QUOTATION
        - ORDER
        - ORDERRESPONSESIMPLE
        - REMITTANCEADVICE
        - INVOICE
        - APPLICATIONRESPONSE
        - TRANSPORTATIONSTATUS
        - DESPATCHADVICE
        - RECEIPTADVICE
        - PPAPREQUEST
        - PPAPRESPONSE
        - TRANSPORTEXECUTIONPLANREQUEST
        - TRANSPORTEXECUTIONPLAN
        - ITEMINFORMATIONREQUEST
        - ITEMINFORMATIONRESPONSE
        - OTHER
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
      status:
        type: string
        enum:
        - APPROVED
        - WAITINGRESPONSE
        - DENIED
        - PASTDUE
      processInstanceID:
        type: string
      initiatorID:
        type: string
      responderID:
        type: string
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
      type:
        type: string
      message:
        type: string